Synch with grub2 cvs (2008-04-05)
git-svn-id: https://svn.aros.org/svn/aros/trunk@28240 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/common/boot/grub2/disk/ata.c
@@ -615,7 +615,7 @@ grub_ata_open (const char *name, grub_disk_t disk)
   else
     disk->total_sectors = dev->size;
 
-  disk->id = (int) dev;
+  disk->id = (unsigned long) dev;
   
   disk->has_partitions = !dev->atapi;
   disk->data = dev;
