+++ arch/common/boot/grub2/disk/host.c
@@ -41,7 +41,7 @@ grub_host_open (const char *name, grub_disk_t disk)
       return grub_error (GRUB_ERR_UNKNOWN_DEVICE, "not a host disk");
 
   disk->total_sectors = 0;
-  disk->id = (int) "host";
+  disk->id = (unsigned long) "host";
   
   disk->has_partitions = 0;
   disk->data = 0;
